+++ schemachange/cli.py
@@ -7,13 +7,14 @@ import time
 import hashlib
 import snowflake.connector
 import warnings
+from pandas import DataFrame
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives.asymmetric import dsa
 from cryptography.hazmat.primitives import serialization
 
 # Set a few global variables here
-_schemachange_version = '2.9.4'
+_schemachange_version = '3.0.0'
 _metadata_database_name = 'METADATA'
 _metadata_schema_name = 'SCHEMACHANGE'
 _metadata_table_name = 'CHANGE_HISTORY'
@@ -89,6 +90,7 @@ def schemachange(root_folder, snowflake_account, snowflake_user, snowflake_role,
   change_history = None
   if (dry_run and change_history_metadata) or not dry_run:
     change_history = fetch_change_history(change_history_table, snowflake_session_parameters, autocommit, verbose)
+    r_scripts_checksum = fetch_r_scripts_checksum(change_history_table, snowflake_session_parameters, autocommit, verbose)
 
   if change_history:
     max_published_version = change_history[0]
@@ -102,7 +104,8 @@ def schemachange(root_folder, snowflake_account, snowflake_user, snowflake_role,
   all_script_names = list(all_scripts.keys())
   # Sort scripts such that versioned scripts get applied first and then the repeatable ones.
   all_script_names_sorted =   sorted_alphanumeric([script for script in all_script_names if script[0] == 'V']) \
-                            + sorted_alphanumeric([script for script in all_script_names if script[0] == 'R'])
+                            + sorted_alphanumeric([script for script in all_script_names if script[0] == 'R']) \
+                            + sorted_alphanumeric([script for script in all_script_names if script[0] == 'A'])
 
   # Loop through each script in order and apply any required changes
   for script_name in all_script_names_sorted:
@@ -115,6 +118,29 @@ def schemachange(root_folder, snowflake_account, snowflake_user, snowflake_role,
         print("Skipping change script %s because it's older than the most recently applied change (%s)" % (script['script_name'], max_published_version))
       scripts_skipped += 1
       continue
+    
+    # Apply only R scripts where the checksum changed compared to the last execution of snowchange
+    if script_name[0] == 'R':
+      # get checksum of current R script
+      with open(script['script_full_path'],'r') as content_file:
+        content = content_file.read().strip()
+        content = content[:-1] if content.endswith(';') else content        
+        # Replace any variables used in the script content
+        content = replace_variables_references(content, vars, verbose)
+        checksum_current = hashlib.sha224(content.encode('utf-8')).hexdigest()
+
+        # check if R file was already executed
+        if script_name in list(r_scripts_checksum['script_name']):
+          checksum_last = list(r_scripts_checksum.loc[r_scripts_checksum['script_name'] == script_name, 'checksum'])[0]
+        else:
+          checksum_last = ''
+
+        # check if there is a change of the checksum in the script
+        if checksum_current == checksum_last:
+          if verbose:
+            print(f"Skipping change script {script_name} because there is no change since the last execution")
+          scripts_skipped += 1
+          continue
 
     print("Applying change script %s" % script['script_name'])
     if not dry_run:
@@ -147,6 +173,7 @@ def get_all_scripts_recursively(root_directory, verbose):
       file_full_path = os.path.join(directory_path, file_name)
       script_name_parts = re.search(r'^([V])(.+)__(.+)\.(?:sql|SQL)$', file_name.strip())
       repeatable_script_name_parts = re.search(r'^([R])__(.+)\.(?:sql|SQL)$', file_name.strip())
+      always_script_name_parts = re.search(r'^([A])__(.+)\.(?:sql|SQL)$', file_name.strip())
 
       # Set script type depending on whether it matches the versioned file naming format
       if script_name_parts is not None:
@@ -157,6 +184,10 @@ def get_all_scripts_recursively(root_directory, verbose):
         script_type = 'R'
         if verbose:
           print("Repeatable file " + file_full_path)
+      elif always_script_name_parts is not None:
+        script_type = 'A'
+        if verbose:
+          print("Always file " + file_full_path)
       else:
         if verbose:
           print("Ignoring non-change file " + file_full_path)
@@ -167,8 +198,14 @@ def get_all_scripts_recursively(root_directory, verbose):
       script['script_name'] = file_name
       script['script_full_path'] = file_full_path
       script['script_type'] = script_type
-      script['script_version'] = None if script_type == 'R' else script_name_parts.group(2)
-      script['script_description'] = (repeatable_script_name_parts.group(2) if script_type == 'R' else script_name_parts.group(3)).replace('_', ' ').capitalize()
+      script['script_version'] = None if script_type in ['R', 'A'] else script_name_parts.group(2)
+      if script_type == 'R':
+        script['script_description'] = repeatable_script_name_parts.group(2).replace('_', ' ').capitalize()
+      elif script_type == 'A':
+        script['script_description'] = always_script_name_parts.group(2).replace('_', ' ').capitalize()
+      else:
+        script['script_description'] = script_name_parts.group(3).replace('_', ' ').capitalize()
+      
       all_files[file_name] = script
 
       # Throw an error if the same version exists more than once
@@ -316,6 +353,25 @@ def create_change_history_table_if_missing(change_history_table, snowflake_sessi
   query = "CREATE TABLE IF NOT EXISTS {0}.{1} (VERSION VARCHAR, DESCRIPTION VARCHAR, SCRIPT VARCHAR, SCRIPT_TYPE VARCHAR, CHECKSUM VARCHAR, EXECUTION_TIME NUMBER, STATUS VARCHAR, INSTALLED_BY VARCHAR, INSTALLED_ON TIMESTAMP_LTZ)".format(change_history_table['schema_name'], change_history_table['table_name'])
   execute_snowflake_query(change_history_table['database_name'], query, snowflake_session_parameters, autocommit, verbose)
 
+def fetch_r_scripts_checksum(change_history_table, snowflake_session_parameters, autocommit, verbose):
+  query = f"SELECT DISTINCT SCRIPT, FIRST_VALUE(CHECKSUM) OVER (PARTITION BY SCRIPT ORDER BY INSTALLED_ON DESC) \
+          FROM {change_history_table['schema_name']}.{change_history_table['table_name']} \
+          WHERE SCRIPT_TYPE = 'R' AND STATUS = 'Success'"
+  results = execute_snowflake_query(change_history_table['database_name'], query, snowflake_session_parameters, autocommit, verbose)
+
+  # Collect all the results into a dict
+  d_script_checksum = DataFrame(columns=['script_name', 'checksum'])
+  script_names = []
+  checksums = []
+  for cursor in results:
+    for row in cursor:
+      script_names.append(row[0])
+      checksums.append(row[1])
+
+  d_script_checksum['script_name'] = script_names
+  d_script_checksum['checksum'] = checksums
+  return d_script_checksum
+
 def fetch_change_history(change_history_table, snowflake_session_parameters, autocommit, verbose):
   query = "SELECT VERSION FROM {0}.{1} WHERE SCRIPT_TYPE = 'V' ORDER BY INSTALLED_ON DESC LIMIT 1".format(change_history_table['schema_name'], change_history_table['table_name'])
   results = execute_snowflake_query(change_history_table['database_name'], query, snowflake_session_parameters, autocommit, verbose)
@@ -334,13 +390,13 @@ def apply_change_script(script, vars, default_database, change_history_table, sn
     content = content_file.read().strip()
     content = content[:-1] if content.endswith(';') else content
 
+  # Replace any variables used in the script content
+  content = replace_variables_references(content, vars, verbose)
+
   # Define a few other change related variables
   checksum = hashlib.sha224(content.encode('utf-8')).hexdigest()
   execution_time = 0
   status = 'Success'
-
-  # Replace any variables used in the script content
-  content = replace_variables_references(content, vars, verbose)
 
   # Execute the contents of the script
   if len(content) > 0: